Made QuickView a plugin.
