+++ src/plugins/quickview/index.ts
@@ -1,0 +1,3 @@
+export * from './src/QuickView.ts';
+export * from './src/QuickViewDefaultSettings.ts';
+export * from './types/index.ts';
