+++ src/plugins/quickview/package.json
@@ -1,0 +1,6 @@
+{
+    "name": "quickview",
+    "description": "",
+    "type": "module",
+    "private": true
+}
