+++ src/plugins/quickview/types/QuickViewSettings.ts
@@ -1,0 +1,5 @@
+import { DIVESettings } from '@shopware-ag/dive';
+
+export type QuickViewSettings = DIVESettings & {
+    lightIntensity: number;
+};
